use global instead of wrapper obj
problem: initializing wrapper.new() in plugin/poplar.vim negated
benefits of autoloading
solution: reduce wrapper.vim to a set of functions, since there was no
good reason it was a class. in particular, Run() now creates a global
dict containing the poplar windows if the dict is not found. this allows
for lazy eval on startup.

<cr> on a file now executes `:drop {filename}`
also, renamed wrapper.vim to poplar.vim for consistency